pyproject.toml
[project]
name = "recommendation-system"
version = "1.0.0"
description = "Hybrid VAE Recommendation System with SBERT embeddings"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Aymane Nouhail"}
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "sentence-transformers>=2.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[tool.pyright]
# Pylance/Pyright configuration
pythonVersion = "3.10"
typeCheckingMode = "basic"  # Options: "off", "basic", "standard", "strict"

# Ignore type errors that won't crash at runtime
reportGeneralTypeIssues = "warning"
reportOptionalMemberAccess = "none"
reportOptionalSubscript = "none"
reportOptionalIterable = "none"
reportArgumentType = "none"
reportAssignmentType = "none"
reportAttributeAccessIssue = "none"
reportIndexIssue = "none"
reportCallIssue = "none"

# These are usually fine with dynamic typing
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
reportUnknownVariableType = "none"
reportMissingTypeStubs = "none"
reportPrivateUsage = "none"

[tool.ruff]
# Ruff linter configuration
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip without explicit strict
]

[tool.ruff.lint.per-file-ignores]
"notebooks/*" = ["F403", "F405"]  # Allow star imports in notebooks

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
